Create SetupNotes
Place this folder at the root of your setup, then add the remaining scripts inside of this directory. The scripts use a path to reach and exicute the functions used to automate the setup. They also store temp-backups and logs to allow you easily restore any changes and manage any error handling easily. The logs are in the .gitignore to protect any potentially sensitive information. When setup is finished, this file can be removed to save space if needed. Just run sudo rm -r /SetupNotes or sudo rm -r ~/SetupNotes to delete it.